Add start command and segregate controller to extensions
