+++ src/main/kotlin/ru/sber/kotlin/school/telegram/bot/model/Dictionary.kt
@@ -26,4 +26,21 @@ class Dictionary(
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "owner_id")
     val owner: User
-)
+
+
+) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as Dictionary
+
+        if (id != other.id) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        return id.hashCode()
+    }
+}
