+++ src/main/kotlin/ru/sber/kotlin/school/telegram/bot/repository/DictionaryRepository.kt
@@ -5,14 +5,21 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 import org.springframework.stereotype.Repository
 import ru.sber.kotlin.school.telegram.bot.model.Dictionary
+import ru.sber.kotlin.school.telegram.bot.model.User
 import java.util.Optional
 
 @Repository
 interface DictionaryRepository : JpaRepository<Dictionary, Long> {
 
     @Query(
-        "select id from dictionary where (owner_id=1 or owner_id=:user) and name=:name",
+        "select * from dictionary where (owner_id=1 or owner_id=:user) and name=:name",
         nativeQuery = true
     )
-    fun findIdByNameForUser(@Param("name") name: String, @Param("user") userId: Long): Optional<Long>
+    fun findByNameForUser(@Param("name") name: String, @Param("user") userId: Long): Optional<Dictionary>
+
+    @Query(
+        "SELECT * FROM dictionary WHERE owner_id = :userId Or owner_id = 1",
+        nativeQuery = true
+    )
+    fun findAllDictionariesByUser(@Param("user") userId: Long): List<Dictionary>
 }
