+++ src/main/kotlin/ru/sber/kotlin/school/telegram/bot/service/TrainingService.kt
@@ -1,6 +1,7 @@
 package ru.sber.kotlin.school.telegram.bot.service
 
 import org.springframework.stereotype.Service
+import org.telegram.abilitybots.api.util.AbilityUtils
 import org.telegram.abilitybots.api.util.AbilityUtils.getChatId
 import org.telegram.abilitybots.api.util.AbilityUtils.getUser
 import org.telegram.telegrambots.meta.api.methods.AnswerInlineQuery
@@ -10,6 +11,7 @@ import org.telegram.telegrambots.meta.api.objects.inlinequery.inputmessageconten
 import org.telegram.telegrambots.meta.api.objects.inlinequery.result.InlineQueryResultArticle
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardMarkup
+import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardRemove
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardButton
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardRow
@@ -21,6 +23,7 @@ import ru.sber.kotlin.school.telegram.bot.repository.BotRedisRepository
 import ru.sber.kotlin.school.telegram.bot.repository.DictionaryRepository
 import ru.sber.kotlin.school.telegram.bot.repository.UserRepository
 import ru.sber.kotlin.school.telegram.bot.util.Button
+import java.util.Collections
 
 @Service
 class TrainingService(
@@ -30,47 +33,12 @@ class TrainingService(
     private val gameSelector: GameSelector
 ) {
 
-    fun getAllFromRedis(upd: Update): SendMessage {
-        val userId = getUser(upd).id
-        val state = botRedisRepository.getState(userId)
-        val dic = botRedisRepository.getDictionary(userId)
-        val style = botRedisRepository.getStyle(userId)
-        val queueSize = botRedisRepository.getQueueSize(userId)
-        val botMsg = botRedisRepository.getBotMsg(userId)
-        val userMsg = botRedisRepository.getUserMsg(userId)
-        val answer = botRedisRepository.getAnswer(userId)
 
-        return SendMessage(
-            getChatId(upd).toString(), "Этап: $state\nСловарь: $dic\nТип игры: $style\nСлов: $queueSize" +
-                    "\nОт бота: $botMsg\nОт юзера: $userMsg\nОтвет: $answer"
-        )
-    }
-
-    fun mainMenu(upd: Update): SendMessage =
-        SendMessage(getChatId(upd).toString(), "Выберите пункт меню").apply {
-            replyMarkup = InlineKeyboardMarkup().apply {
-                clearByUser(getUser(upd).id)
-                val trainButton = InlineKeyboardButton("Тренировка")
-                    .also {
-                        it.switchInlineQueryCurrentChat = "dicts"
-                    }
-                keyboard = listOf(listOf(trainButton))
-            }
-        }
-
-    private fun clearByUser(userId: Long) {
-        botRedisRepository.clearQueue(userId)
-        botRedisRepository.deleteBotMsg(userId)
-        botRedisRepository.deleteUserMsg(userId)
-        botRedisRepository.deleteAnswer(userId)
-        botRedisRepository.deleteDictionary(userId)
-        botRedisRepository.deleteStyle(userId)
-    }
 
     fun getFavorites(upd: Update): AnswerInlineQuery {
         val query = upd.inlineQuery
 
-        val user: User = userRepository.findById(1)
+        val user: User = userRepository.findById(getUser(upd).id)
             .orElse(User(1, "admin", "admin", "admin"))
         val results = user.favorites
             .map { prepareDictArticle(it) }
@@ -107,7 +75,7 @@ class TrainingService(
         .build().also {
             val dictionaryName = upd.message.text.substringAfter('\n')
             val userId = getUser(upd).id
-            val dictionaryId = dictionaryRepository.findIdByNameForUser(dictionaryName, userId).get()
+            val dictionaryId = dictionaryRepository.findByNameForUser(dictionaryName, userId).get().id
             botRedisRepository.putToDictionary(userId, dictionaryId)
         }
